Set voltageChart min y-value to 2V
As mentioned in liamcottle/meshtastic-map#13 , voltages on
meshtastic devices are typically 3.2-4.2V .  The range on the
voltageChart y-axis from 0V to 5V reduced its usefulness as changes
within the normal operating range of devices were more difficult to
detect.

This patch changes the y-axis range to 2V to 5V. This will improve
the resolution for the areas users care about.

Note that this change is more conservative than requested in
liamcottle/meshtastic-map#13 , as it attempts to take into account
less typical voltage. A very drained battery might get around 2V,
and 5V might be reported for a USB connection.